KA-137 - Rename ListOrSpinner files to ListingOrSpinner
