+++ onboarding-task/src/App.jsx
@@ -1,6 +1,7 @@
 import './sticky-footer.css';
 import React, { Component } from 'react';
-import { ListOrSpinner } from './containers/ListOrSpinner.ts';
+import { ListOrSpinner } from './containers/ListingOrSpinner.ts';
+import { List } from './containers/List.ts';
 
 class App extends Component {
   render() {
@@ -23,7 +24,7 @@ class App extends Component {
           </div>
 
           <section id="app-content">
-            <ListOrSpinner />
+            <ListOrSpinner wrappedComponent={<List />} />
           </section>
         </div>
         <footer className="footer">
