+++ onboarding-task/src/components/ListingOrSpinner.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 import * as PropTypes from 'prop-types';
-import { List } from '../containers/List';
 import { IAction } from '../actionCreators/IAction';
 
 export interface IListOrSpinnerDataProps {
   isFetching: boolean;
+  wrappedComponent: any;
 }
 
 export interface  IListOrSpinnerCallbackProps {
@@ -17,6 +17,7 @@ class ListOrSpinner extends React.PureComponent<ListOrSpinnerProps>{
   static displayName = 'ListOrSpinner';
   static propTypes = {
     isFetching: PropTypes.bool.isRequired,
+    wrappedComponent: PropTypes.element.isRequired,
     fetchItems: PropTypes.func.isRequired,
   };
 
@@ -30,7 +31,7 @@ class ListOrSpinner extends React.PureComponent<ListOrSpinnerProps>{
 
   render() {
     return this.props.isFetching ?
-      <img src={require('../../assets/running_spinner.gif')} /> : <List/>;
+      <img src={require('../../assets/running_spinner.gif')} /> : this.props.wrappedComponent;
   }
 }
